+++ CHANGELOG.md
@@ -10,7 +10,8 @@ Polyverse Boost Automatic Application Modernization
 - Added searchable product help documentation to the Boost Activity Bar Chat window
 
 ### Enhancements
-- N/A
+- Disabled by default legal analysis kernels: Single function blueprint, Coding Guidelines, Function Summary
+    - Quick Blueprint and Custom Scan have replaced these for most scenarios
 
 ### Bug Fixes
 - Fix for Analysis Estimate not refreshing after analysis completes